pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [{ name = "Adrian", email = "adrianoopereiira@hotmail.com" }]
#readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "pytest>=8.3.4,<9.0.0",
    "factory-boy>=3.3.1,<4.0.0",
    "django>=5.1.5,<6.0.0",
    "djangorestframework>=3.15.2,<4.0.0",
    "django-extensions>=3.2.3,<4.0.0",
    "django-debug-toolbar>=5.0.1,<6.0.0"
]

[optional-dependencies]
dev = [
    "pytest-django^4.9.0"
]
